0331 kodeal → blog 통합
