+++ templates/common/signup_complete.html
@@ -4,7 +4,7 @@
 
 <div class="container" align="center">
     <font size="4" style="font-weight: 700;"> 회원가입 성공</font>
-    <button type="button" onclick="location.href='{% url 'kodeal:index' %}'"> 홈으로 </button>
+    <button type="button" onclick="location.href='{% url 'blog:index' %}'"> 홈으로 </button>
 </div>
 
 {% endblock %}
