+++ templates/navBar.html
@@ -19,7 +19,7 @@
     <!-- QnA화면 진입 -->
     <div class="nav-item" style="float:right; margin-right: 20px;">
         {% if user.is_authenticated %}
-            <a href="{% url 'kodeal:qna_main' %}"><button>QnA</button></a>
+            <a href="{% url 'blog:qna_main' %}"><button>QnA</button></a>
         {% endif %}
     </div>
     <!-- 사용자 정보 및 로그아웃 -->
